transport/db: disable SEQUENCENUMBERS + widen remote slice get/put to uint64
(1) SEQUENCENUMBERS is a debug-only out-of-order detector that indexes
rec_seq_numbers[packet->seq_rank] in the hot receive path; a mis-framed
large-transfer packet yields a garbage seq_rank and SIGSEGVs in
arts_server_process_packet. Disabled (diagnostic only, untrusted index).
conv-2d 2n now runs to completion (exit 0) instead of crashing.

(2) The local DB-slice path carries uint64 slice_offset/slice_size but the
remote path truncated them to (unsigned int); widen the remote get/put chain
(arts_remote_get_from_db, arts_get_from_db[_at][_ex], internal_get_from_db,
arts_out_of_order_get_from_db, arts_alloc_db_slice_payload, get_put packet,
oo_get_from_db_s) to uint64 and drop the casts. Runtime-only; compiler already
passes uint64 slice_offset. gemm 2n unchanged (1.52x, correct); conv-2d 2n
correct at <=4GB.

+++ libs/include/internal/arts/gas/out_of_order.h
@@ -104,8 +104,8 @@ void arts_out_of_order_handle_remote_db_full_send(arts_guid_t db_guid, int rank,
                                                   unsigned int slot,
                                                   arts_db_access_mode_t mode);
 void arts_out_of_order_get_from_db(arts_guid_t edt_guid, arts_guid_t db_guid,
-                                   unsigned int slot, unsigned int offset,
-                                   unsigned int size, uint32_t flags);
+                                   unsigned int slot, uint64_t offset,
+                                   uint64_t size, uint32_t flags);
 void arts_out_of_order_signal_edt_with_ptr(arts_guid_t edt_guid,
                                            arts_guid_t db_guid, void *ptr,
                                            unsigned int size,
