+++ libs/include/internal/arts/memory/db.h
@@ -91,8 +91,8 @@ void internal_put_in_db(void *ptr, arts_guid_t edt_guid, arts_guid_t db_guid,
                         unsigned int size, arts_guid_t epoch_guid,
                         unsigned int rank);
 void arts_get_from_db_at_ex(arts_guid_t edt_guid, arts_guid_t db_guid,
-                            unsigned int slot, unsigned int offset,
-                            unsigned int len, uint32_t flags,
+                            unsigned int slot, uint64_t offset,
+                            uint64_t len, uint32_t flags,
                             unsigned int rank);
 
 void arts_db_destroy_safe(arts_guid_t guid, bool remote);
